Add comments and break up helper
