+++ app/controllers/registration_controller.rb
@@ -1,17 +1,22 @@
+# RegistrationController
+#
+# Handles incoming registration requests from tool
+# consumers.
 class RegistrationController < ApplicationController
   include RegistrationHelper
 
   def register
     # Fetch the a custom tool consumer profile
-    tcp_helper = ToolConsumerProfileHelper.new(URI.parse(params[:tc_profile_url]), jwt_access_token)
+    tcp_helper = Registration::ToolConsumerProfile.new(URI.parse(params[:tc_profile_url]), jwt_access_token)
 
     # Do failure redirect to tool consumer if missing required
+    # TODO check for require services
     redirect_to registration_failure_url and return unless tcp_helper.supports_required_capabilities?
 
     # Create a new tool proxy
     tool_proxy = ToolProxy.new(tcp_url: tcp_helper.tcp_url, base_url: request.base_url)
     tp_response = register_tool_proxy(tool_proxy, tcp_helper.tp_service_url, jwt_access_token)
-    tp_response_helper = ToolProxyResponseHelper.new(tp_response, registration_redirect_url)
+    tp_response_helper = Registration::ToolProxyResponse.new(tp_response, registration_redirect_url)
 
     # Persist tool proxy and redirect if no errors
     if tp_response_helper.success?
